Update CURD function of star_wars_character to use native SQL query instead of built-ins

+++ src/main/java/com/lds/swapi/model/GalaxyPlanet.java
@@ -8,18 +8,17 @@ public class GalaxyPlanet {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
+    private Integer id;
     private String name;
     private String climate;
     private Integer population;
 
     // Getters and setters
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
